Fixed shop, level updates, musik + kommentarer
Shoppen fungerar nu, man kan köpa upgrades och de fungerar också. Små level changes. Musik spelas nu i Menu och Level scenerna.

+++ Assets/Scripts/PlayerScripts/PlayerMovement.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class PlayerMovement : MonoBehaviour
 {
-    public float walkSpeed;
+    public int walkSpeed;
     public Rigidbody2D rb;
     OpenShop shop;
